made repository methods virtual
this is necessary, because every now and then you need the object to be found with a .include subclass, which is not possible with the regular .find method.  so find is definetly useful to be made as virtual.  the rest of the methods were made virtual as a just in case